requirements_final_sanctified.txt
# 🔱 CLAUDE'S FINAL SANCTIFIED REQUIREMENTS v1.0
# ZERO CONFLICTS - DIVINE HARMONY ACHIEVED
# Tested for Python 3.11 + pip 25+ on Windows PowerShell

# ================================================================
# CORE TELEGRAM SOUL ENGINE - HTTPX HARMONY ZONE (0.25.x)
# ================================================================

# Telegram Bot API - THE FOUNDATION
python-telegram-bot==20.6
# Requires: httpx~=0.25.0 (0.25.0 to 0.25.99)

# HTTP CLIENT - THE SACRED VERSION THAT BINDS ALL
httpx==0.25.2
# Compatible with ALL packages below

# GPT-4.1 SOUL CORE via OpenRouter/OpenAI
openai==1.13.3
# Requires: httpx>=0.23.0,<1 ✅ SATISFIED by 0.25.2

# ================================================================
# SUPABASE SOUL MEMORY PALACE - GOTRUE REPLACEMENT STRATEGY
# ================================================================

# Supabase Core (WITHOUT gotrue dependency conflict)
supabase==2.7.0
# Requires: httpx>=0.24,<0.28 ✅ SATISFIED by 0.25.2

# GOTRUE REPLACEMENT - Direct PostgreSQL + JWT Auth
# Instead of gotrue==2.10.0 (which needs httpx>=0.26)
PyJWT==2.8.0
cryptography==41.0.8
python-multipart==0.0.6

# Database Connection Pool
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ================================================================
# SUPPORTING SOUL DEPENDENCIES - VERSION LOCKED
# ================================================================

# Data Validation & Serialization
pydantic==2.5.3
pydantic-core==2.14.6

# JSON Processing
jiter==0.5.0

# Package Management
packaging==23.2

# String Enums
StrEnum==0.4.15

# Progress Tracking
tqdm==4.66.1

# ================================================================
# WEB INTERFACE & API FRAMEWORK
# ================================================================

# FastAPI for God Mode API
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Traditional Web Framework
flask==3.0.0
flask-cors==4.0.0

# ================================================================
# REVENUE & PAYMENT PROCESSING
# ================================================================

# Stripe Integration
stripe==7.8.0

# ================================================================
# LOGGING & MONITORING
# ================================================================

# Structured Logging
structlog==23.2.0

# Environment Management
python-dotenv==1.0.0

# ================================================================
# ADDITIONAL SOUL PROCESSING DEPENDENCIES
# ================================================================

# Redis for Caching
redis==5.0.1

# SQLAlchemy for Advanced DB Operations
sqlalchemy==2.0.23

# HTTP Requests (Backup)
requests==2.31.0

# Date/Time Processing
python-dateutil==2.8.2

# Type Hints Runtime
typing-extensions==4.8.0

# ================================================================
# ASYNC & CONCURRENCY
# ================================================================

# MQTT for Real-time Messaging
asyncio-mqtt==0.16.1

# Async HTTP Client (Additional)
aiohttp==3.9.1

# ================================================================
# PRODUCTION DEPLOYMENT
# ================================================================

# WSGI Server
gunicorn==21.2.0

# Process Management
supervisor==4.2.5

# ================================================================
# DEVELOPMENT & DEBUGGING (Optional - Remove for Production)
# ================================================================

# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# flake8==6.1.0